+++ .github/workflows/deploy.yml
@@ -57,6 +57,9 @@ jobs:
           tmux new-session -d -s \"\${SESSION_NAME}\" '
             source .venv/bin/activate
             uv pip install -r requirements.txt
+            export APP_VERSION=${TAG}
+            export APP_USERNAME=aiuser
+            export APP_PASSWORD=${{ secrets.APP_PASSWORD }}
             export ANTHROPIC_API_KEY=${{ secrets.ANTHROPIC_API_KEY }}
             export OPENAI_API_KEY=${{ secrets.OPENAI_API_KEY }}
             export POSTGRES_CONNECTION_STRING=${{ secrets.POSTGRES_CONNECTION_STRING }}
