+++ src/utils/auth.py
@@ -1,0 +1,25 @@
+from fastapi import status
+from fastapi.security import  HTTPBasicCredentials, HTTPBasic
+from fastapi import Depends, HTTPException
+from src.constants import APP_PASSWORD, APP_USERNAME
+import secrets
+
+def verify_credentials(credentials: HTTPBasicCredentials = Depends(HTTPBasic())):
+    current_username_bytes = credentials.username.encode("utf8")
+    correct_username_bytes = APP_USERNAME.encode("utf8")
+    is_correct_username = secrets.compare_digest(
+        current_username_bytes, correct_username_bytes
+    )
+    current_password_bytes = credentials.password.encode("utf8")
+    correct_password_bytes = APP_PASSWORD.encode("utf8")
+    is_correct_password = secrets.compare_digest(
+        current_password_bytes, correct_password_bytes
+    )
+    
+    if not (is_correct_username and is_correct_password):
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Incorrect username or password",
+            headers={"WWW-Authenticate": "Basic"},
+        )
+    return credentials.username
